refactor(frontend): lift interfaces from watch to composable
Lift typescript interfaces from watch to useResourceWatch
